`chore:` refactor app structure
