+++ src/assets/styles/DashboardHeader.css
@@ -1,4 +1,4 @@
-@import '../index.css';
+@import '../../index.css';
 
 #dashboard {
 	display: flex;
