+++ src/assets/svgs/Delete.jsx
