+++ src/assets/svgs/Edit.jsx
