+++ src/assets/svgs/Search.jsx
